toolkit/components/satchel/test/test_basic_form.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for satchel</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script> 
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
Satchel test: simple form fill
<p id="display"></p>

<iframe id="iframe"></iframe>

<div id="content" style="display: none">
</div>

<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Satchel **/
function startTest() {
  netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');

  // Get the formHistory service
  var formHistory = Components.classes["@mozilla.org/satchel/form-history;1"].
                    getService(Components.interfaces.nsIFormHistory2);

  // Make sure it's not already initialized somehow
  ok(!formHistory.hasEntries, "no existing entries");

  // add some test entries
  var testData = [{name: "text", value: "data-basic", expect: "data-basic"},
                  {name: "text-max", value: "data-maxlength01234567890", expect: "data"}
                 ];

  testData.forEach(function (entry) formHistory.addEntry(entry.name, entry.value));

  // Simple check to see if everything was added fine.
  ok(formHistory.hasEntries, "entries exist");

  // Now load a document and check that the form was filled out correctly.
  var iframe = $("iframe");
  iframe.addEventListener("load", doTest, false);

  var formDocURL = 'data:text/html,<form><input type="text" id="text" name="text"><input type="text" id="text-max" name="text-max" maxlength="4"></form>'
  iframe.setAttribute("src", formDocURL);

  function doTest() {
    iframe.removeEventListener("load", doTest, false);

    var text = iframe.contentWindow.document.getElementById("text");
    var text_max = iframe.contentWindow.document.getElementById("text-max");
    ok(text, "got text element");
    ok(text_max, "got text-max element");

    function triggerFormFill(aTarget) {
      aTarget.focus();
      synthesizeKey("VK_DOWN", {}, window);
      synthesizeKey("VK_DOWN", {}, window);
      synthesizeKey("VK_ENTER", {}, window);
    }

    triggerFormFill(text);
    triggerFormFill(text_max);

    is(text.value, testData[0].value, "text filled in correctly");
    is(text_max.value, testData[1].value, "text_max filled in correctly");

    // clean up
    cleanUp();
  }
  
  function cleanUp() {
    netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');

    // Remove the added entries
    formHistory.removeAllEntries();

    SimpleTest.finish();
  }
}

window.onload = startTest;

SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
